Make UI tests use the test config in config.ini.php
The command that runs UI tests now writes a /tests/UI/config.js file that contains the variable configured in config.ini.php. If a file already exists and wasn't autogenerated by this command, it is not overwritten.

That allows to run UI tests with another remote than localhost, which happens for example with Docker where phantomjs runs in another container than Piwik.