Docker: leave the tmp directory where it is and instead mount it as a local volume
Moving Piwik's tmp directory to /tmp/piwik (or something similar) doesn't work because the tmp path is hardcoded in Piwik -> we have to leave it inside Piwik's directory.

I now mount a volume to replace the tmp/ directory -> this volume is empty and local to the container. But it has root permissions. To solve that, everytime the data container is started a chmod is run (else permissions are lost on each container run).